Generate a Grid and Calculate PPI Across a Marine Study Area

Creates a regular spatial grid across a user-specified bounding box with an added buffer. Excludes fully terrestrial grid cells based on land polygon overlap. For the remaining cells (partially or fully marine), it calculates Population Pressure Index (PPI) at one or more buffer distances.

Usage

generate_ppi_grid(
  lon,
  lat,
  buffer_km = c(5, 10, 20),
  tile_schema,
  data_directory,
  land_polygons,
  grid_resolution_km = 1,
  locations_crs = 4326,
  crs_projected = 3857
)

Arguments

lon

Numeric vector of length 1 or greater: longitudes in decimal degrees representing the range (or centroid) of the study area.

lat

Numeric vector of length 2: latitudes in decimal degrees representing the range (or centroid) of the study area.

buffer_km

Numeric vector of one or more buffer radii (in kilometers) used to estimate PPI.

tile_schema

An sf object of GHSL tile polygons with a tile_id column

data_directory

The base directory containing "downloads" and "mosaics" folders

land_polygons

An sf object representing land areas (e.g., from Natural Earth or OpenStreetMap).

grid_resolution_km

Numeric. Width and height of grid cells in kilometers.

locations_crs

CRS code for the input coordinates (default is 4326)

crs_projected

CRS to use for grid generation and area calculations (default: 3857 Web Mercator).

Value

A data.frame of grid centroids and PPI estimates at each buffer distance.